+++ src/close.rs
@@ -1,9 +1,8 @@
 use crate::close::close_xml::do_closing_xml;
-use crate::close::close_xml::Budget;
+use crate::close::close_xml::read_budget_config;
 use crate::derive_month_from_accounts;
 
 use std::error::Error;
-use std::fs;
 use std::path::Path;
 
 mod close_xml;
@@ -26,29 +25,5 @@ pub fn close(
         }
     } else {
         Err("No valid file extension found".into())
-    }
-}
-
-fn read_budget_config(budget_config_file: &Path) -> Result<Budget, Box<dyn Error>> {
-    let toml_str = fs::read_to_string(budget_config_file)?;
-    let budget: Budget = toml::from_str(&toml_str)?;
-    Ok(budget)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use rstest::rstest;
-    use std::path::PathBuf;
-
-    #[rstest]
-    fn can_read_from_sample_config_file() {
-        let file_path = "sample_config/budget.toml".to_string();
-        let config_file = &PathBuf::from(file_path);
-        let config = read_budget_config(config_file);
-        let Ok(config) = config else {
-            panic!("Invalid config: {config:#?}");
-        };
-        assert_eq!("LoLa Budget", config._name);
     }
 }
